Refactor: Exchange old flowView
