+++ src/views/FlowView.tsx
@@ -1,14 +1,7 @@
-import React, { useEffect, useState, useCallback } from "react"
-import { useSelector, useDispatch } from "react-redux"
-import {
-  selectFlowArguments,
-  selectFlowChart,
-  selectFlows,
-  selectRerender,
-  selectSelectedFlowId,
-  selectTooltipConfig,
-} from "../redux/flows/flows.selectors"
-import * as actions from "@bastinjafari/react-flow-chart-with-tooltips-and-multi-select/src/container/actions"
+import React, { useCallback, useState } from "react"
+import { Card, Container, Row } from "shards-react"
+import { PageTitle } from "../components/Common/PageTitle"
+import FlowSelection from "../components/FlowChart/FlowSelection"
 import {
   createNewFlow,
   deleteFlow,
@@ -16,28 +9,25 @@ import {
   loadFlow,
   startFlow,
   stopFlow,
-  updateFlow,
 } from "../redux/flows/flows.actions"
-import html2canvas from "html2canvas"
-import { cloneDeep } from "lodash"
-import { copyToClipboard, formatAsYAML } from "../helpers"
-import { Card, Container, Row } from "shards-react"
-import { PageTitle } from "../components/Common/PageTitle"
-import FlowSelection from "../components/FlowChart/FlowSelection"
 import CommandBar from "../components/FlowChart/CommandBar"
-import {
-  FlowChart,
-  IOnLinkCompleteInput,
-} from "@bastinjafari/react-flow-chart-with-tooltips-and-multi-select"
-import Tooltip from "../components/FlowChart/Tooltip"
-import CustomNode from "../components/FlowChart/ChartNode"
-import CustomPort from "../components/FlowChart/NodePort"
-import Sidebar from "../components/FlowChart/Sidebar"
-import { showModal } from "../redux/global/global.actions"
-import { selectConnectionStatus } from "../redux/global/global.selectors"
 
+import Sidebar from "../components/FlowChart/Sidebar"
 import styled from "@emotion/styled"
+import { useDispatch, useSelector } from "react-redux"
+import {
+  selectFlowArguments,
+  selectFlow,
+  selectFlows,
+  selectRerender,
+  selectSelectedFlowId,
+} from "../redux/flows/flows.selectors"
+import { selectConnectionStatus } from "../redux/global/global.selectors"
+import { showModal } from "../redux/global/global.actions"
 import logger from "../logger"
+import { copyToClipboard, formatAsYAML } from "../helpers"
+import html2canvas from "html2canvas"
+import ReactFlow from "react-flow-renderer"
 
 const FlowViewContainer = styled.div`
   display: flex;
@@ -46,7 +36,6 @@ const FlowViewContainer = styled.div`
 const FlowContainer = styled.div`
   overflow: hidden;
 `
-
 export default function FlowView() {
   const dispatch = useDispatch()
   useState(useSelector(selectRerender))
@@ -54,31 +43,19 @@ export default function FlowView() {
   const selectedFlowId = useSelector(selectSelectedFlowId)
   const flows = useSelector(selectFlows)
   const flowArguments = useSelector(selectFlowArguments)
-  const flowChart = useSelector(selectFlowChart)
-  const { flow: chart, type: flowType } = flowChart
-  const chartWithTooltips = {
-    ...chart,
-    ...useSelector(selectTooltipConfig),
-  }
+  const flow = useSelector(selectFlow)
+  const {
+    flowChart: { elements },
+    type: flowType,
+  } = flow
+
+  const copyChartAsYAML = useCallback(() => {
+    logger.log("copyChartAsYAML | chart:", elements)
+    copyToClipboard(formatAsYAML(elements, flowArguments))
+    alert("Chart copied to clipboard as YAML")
+  }, [elements, flowArguments])
+
   const [showOverlay, setShowOverlay] = useState<boolean>(false)
-
-  const actionCallbacks = Object.keys(actions).reduce((obj: any, key: any) => {
-    obj[key] = (...args: any) => {
-      let action = (actions as any)[key]
-      let newChartTransformer = action(...args)
-      const chartCopy = cloneDeep(chart) //todo this is necessary because our state management is not consistent. Remove with new flow library
-      let newChart = newChartTransformer(chartCopy)
-      dispatch(updateFlow({ ...chartWithTooltips, ...newChart }))
-      return newChart
-    }
-    return obj
-  }, {})
-
-  useEffect(() => {
-    const chartContainer = document.querySelector(".chart-container")
-    if (chartContainer)
-      chartContainer.addEventListener("contextmenu", (e) => e.preventDefault())
-  }, [])
 
   const showCaptureOverlay = (showOverlay = true) => {
     setShowOverlay(showOverlay)
@@ -113,25 +90,31 @@ export default function FlowView() {
     })
   }
 
+  const handleDuplicateFlow = () => {
+    const flowYAML = formatAsYAML(elements, flowArguments)
+    dispatch(duplicateFlow(flowYAML))
+  }
+
+  //todo remove any
   const updateNode = (node: any) => {
-    let newChart = cloneDeep(chart)
-    newChart.nodes[node.id].label = node.label
-
-    let props = {
-      ...node.properties,
-      ...node.newProperties,
-    }
-
-    Object.keys(props).forEach((id) => {
-      if (props[id] === "" || typeof props[id] === "undefined") {
-        delete props[id]
-      }
-    })
-
-    newChart.nodes[node.id].properties = props
-    dispatch(updateFlow({ ...chart, ...newChart }))
-
-    return newChart.nodes[node.id]
+    // let newChart = cloneDeep(chart)
+    // newChart.elements[node.id].label = node.label
+    //
+    // let props = {
+    //   ...node.properties,
+    //   ...node.newProperties,
+    // }
+    //
+    // Object.keys(props).forEach((id) => {
+    //   if (props[id] === "" || typeof props[id] === "undefined") {
+    //     delete props[id]
+    //   }
+    // })
+    //
+    // newChart.nodes[node.id].properties = props
+    // dispatch(updateFlow({ ...chart, ...newChart }))
+    //
+    // return newChart.nodes[node.id]
   }
 
   const updateLink = (
@@ -139,70 +122,14 @@ export default function FlowView() {
     fromId: string,
     toId: string | undefined
   ) => {
-    if (fromId === toId) return
-    let newChart = cloneDeep(chart)
-
-    newChart.links[linkId].from.nodeId = fromId
-    newChart.links[linkId].to.nodeId = toId
-
-    dispatch(updateFlow({ ...chart, ...newChart }))
+    // if (fromId === toId) return
+    // let newChart = cloneDeep(chart)
+    //
+    // newChart.links[linkId].from.nodeId = fromId
+    // newChart.links[linkId].to.nodeId = toId
+    //
+    // dispatch(updateFlow({ ...chart, ...newChart }))
   }
-
-  const deleteSelection = () => {
-    actionCallbacks.onDeleteKey({})
-  }
-
-  const copyChartAsYAML = useCallback(() => {
-    logger.log("copyChartAsYAML | chart:", chart)
-    copyToClipboard(formatAsYAML(chart, flowArguments))
-    alert("Chart copied to clipboard as YAML")
-  }, [chart, flowArguments])
-
-  const validateLink = ({
-    fromNodeId,
-    toNodeId,
-    fromPortId,
-    toPortId,
-  }: IOnLinkCompleteInput) => {
-    return !(fromNodeId === toNodeId || fromPortId === toPortId)
-  }
-
-  const showImportModal = () => {
-    dispatch(showModal("import"))
-  }
-
-  const showNewFlowModal = () => {
-    dispatch(showModal("newFlow"))
-  }
-
-  const handleCreateNewFlow = (e: any) => {
-    e.preventDefault()
-    e.stopPropagation()
-
-    dispatch(createNewFlow())
-  }
-
-  const handleDeleteFlow = (e: any, flowId: any) => {
-    e.preventDefault()
-    e.stopPropagation()
-
-    dispatch(deleteFlow(flowId))
-  }
-
-  const handleDuplicateFlow = () => {
-    const flowYAML = formatAsYAML(chart, flowArguments)
-    dispatch(duplicateFlow(flowYAML))
-  }
-
-  const handleStartFlow = () => {
-    dispatch(startFlow(selectedFlowId))
-  }
-
-  const handleStopFlow = () => {
-    dispatch(stopFlow(selectedFlowId))
-  }
-
-  const readonly = flowType !== "user-generated"
 
   return (
     <Container fluid className="main-content-container px-0">
@@ -211,7 +138,7 @@ export default function FlowView() {
           download
         </a>
         <Row noGutters className="page-header mb-4">
-          <PageTitle title="Flow Design" className="text-sm-left mb-3" />
+          <PageTitle title="FlowChart Design" className="text-sm-left mb-3" />
         </Row>
 
         <FlowViewContainer>
@@ -219,19 +146,19 @@ export default function FlowView() {
             connected={connected}
             flows={flows}
             selectedFlowId={selectedFlowId}
-            showNewFlowModal={showNewFlowModal}
-            createNewFlow={handleCreateNewFlow}
+            showNewFlowModal={() => dispatch(showModal("newFlow"))}
+            createNewFlow={() => dispatch(createNewFlow())}
             loadFlow={(flowId) => dispatch(loadFlow(flowId))}
-            deleteFlow={handleDeleteFlow}
+            deleteFlow={(e, flowId) => dispatch(deleteFlow(flowId))}
           />
 
           <FlowContainer>
             <Card className="chart-section-container mr-md-4 mb-4">
               <CommandBar
-                startFlow={handleStartFlow}
-                stopFlow={handleStopFlow}
+                startFlow={() => dispatch(startFlow(selectedFlowId))}
+                stopFlow={() => dispatch(stopFlow(selectedFlowId))}
                 copyChart={copyChartAsYAML}
-                importChart={showImportModal}
+                importChart={() => dispatch(showModal("import"))}
                 exportImage={exportImage}
               />
               <div className="chart-container">
@@ -242,20 +169,10 @@ export default function FlowView() {
                   <div className="capture-overlay-top" />
                   <div className="capture-overlay-bottom" />
                 </div>
+              </div>
 
-                <FlowChart
-                  Components={{
-                    TooltipComponent: Tooltip,
-                    NodeInner: CustomNode as any,
-                    Port: CustomPort,
-                  }}
-                  callbacks={actionCallbacks}
-                  config={{
-                    readonly,
-                    validateLink: validateLink,
-                    smartRouting: true,
-                  }}
-                />
+              <div style={{ height: 1000, width: 1000 }}>
+                <ReactFlow elements={elements} />
               </div>
             </Card>
           </FlowContainer>
@@ -263,9 +180,9 @@ export default function FlowView() {
           <Sidebar
             arguments={flowArguments.pod}
             duplicateFlow={handleDuplicateFlow}
-            readonly={readonly}
-            flow={chart}
-            deleteSelection={deleteSelection}
+            readonly={flowType !== "user-generated"}
+            elements={elements}
+            deleteSelection={() => {}}
             updateNode={updateNode}
             updateLink={updateLink}
           />
